Add instructions and move country codes to json
